removed the markdown file
