+++ app/data-actions/milestones.js
@@ -1,9 +1,27 @@
 import moment from 'moment'
 import schedule from 'node-schedule'
-import * as markdown from './../markdown'
 import data from './../../mock_data/milestones'
 
-// ---------------------- milestones ---------------------- //
+/*
+ * Takes a given header, makes it bold, all uppercase, and adds a
+ * newline character at the ned of it.
+ */
+const formatHeader = (key) => {
+  return `\n*${key.toUpperCase()}*:\n`
+}
+
+/*
+ * Format items in a dictionary into a bullet list format where each
+ * line is tabbed.
+ */
+const formatLists = (dict) => {
+  let result = ''
+
+  for (const item in dict) {
+    result += `\t• ${dict[item]}\n`
+  }
+  return result
+}
 
 /*
 * getMilestone() takes in the current week number and gets and formats the milestone for
@@ -18,10 +36,10 @@ export const getMilestone = (week) => {
 
   // formats each section of the milestone in markdown format
   for (const section in rawMilestone) {
-    if (section != 'milestone' && section != 'title') {
-      milestone += markdown.formatHeader(section)
+    if (section !== 'milestone' && section !== 'title') {
+      milestone += formatHeader(section)
       if (typeof rawMilestone[section] === typeof new Object()) {
-        milestone += markdown.formatLists(rawMilestone[section])
+        milestone += formatLists(rawMilestone[section])
       } else {
         milestone += (`${rawMilestone[section]}\n`)
       }
@@ -33,31 +51,3 @@ export const getMilestone = (week) => {
 
   return milestone
 }
-
-// start and end dates for term and week count
-const startDate = moment('03/28/2018', 'MM/DD/YYYY')
-const endDate = moment('05/30/2018', 'MM/DD/YYYY')
-let currentWeek = 0
-
-// schedule milestone message for Wednesday after the Lab meeting at 8pm
-const rule = new schedule.RecurrenceRule()
-rule.dayOfWeek = 3
-rule.hour = 20
-rule.minute = 0
-rule.second = 0
-
-
-
-// // if the current date is within the start and end date range
-// if (moment().range(startDate, endDate) && currentWeek < 10) {
-//   // post milestones every Tuesday at 10AM
-//   schedule.scheduleJob({ hour: 10, dayOfWeek: 2 }, () => {
-//     channels.forEach(channel => {
-//       slackbot.say({
-//         text: getMilestone(currentWeek),
-//         channel,
-//       })
-//       currentWeek += 1
-//     })
-//   })
-// }
